dockerExploration/FastAPIDocker/requirements.txt
fastapi
uvicorn
sqlalchemy
async-exit-stack
async-generator
passlib[bcrypt]
python-jose[cryptography]
python-multipart
python-dotenv
mysql-connector-python
pymysql